+++ front-end/src/App.js
@@ -10,6 +10,7 @@ import Checkout from './pages/Checkout';
 import PagesOrderDetails from './pages/PagesOrderDetails';
 import Register from './pages/Register';
 import PageSellerProducts from './pages/PageSellerProducts';
+import PagesSellerOrderDetails from './pages/PagesSellerOrderDetails';
 
 function App() {
   return (
@@ -26,6 +27,7 @@ function App() {
           <Route exact path="/customer/orders" component={ PagesOrder } />
           <Route exact path="/customer/orders/:id" component={ PagesOrderDetails } />
           <Route exact path="/seller/products" component={ PageSellerProducts } />
+          <Route exact path="/seller/orders/:id" component={ PagesSellerOrderDetails } />
         </Switch>
       </AppProvider>
     </div>
