+++ front-end/src/components/DetailsSellerOrder.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function DetailsSellerOrder() {
+  return (
+    <div />
+  );
+}
