+++ front-end/src/pages/PagesSellerOrderDetails.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function PagesSellerOrderDetails() {
+  return (
+    <div />
+  );
+}
